CGA-Harvester/requirements.txt
kafka_python==1.3.1
lz4==0.8.2
lz4tools==1.3.1.2
xxhash==0.6.1
kafka==1.3.1
msgpack_python==0.4.8
twitter==1.17.1
